test: let mocked warehouse event tests run in the CI
